Extract exploration state management
