+++ assets/js/dashboard/extra/exploration/constants.ts
@@ -1,0 +1,11 @@
+export const DIRECTION = { FORWARD: 'forward', BACKWARD: 'backward' }
+
+export const DIRECTION_OPTIONS = [
+  { value: DIRECTION.FORWARD, label: 'Starting point' },
+  { value: DIRECTION.BACKWARD, label: 'End point' }
+]
+
+export const PAGE_FILTER_KEYS = ['page', 'entry_page', 'exit_page']
+
+export const MAX_VISIBLE_CANDIDATES = 10
+export const MIN_GRID_COLUMNS = 3
